refactor: move homepage component to templates file
