+++ components/templates/Home/FAQSection/FAQBox.tsx
