+++ components/templates/Home/FAQSection/FAQSection.module.scss
