+++ components/templates/Home/FAQSection/FAQSection.tsx
