+++ components/templates/Home/FAQSection/index.ts
