+++ components/templates/Home/IntroSection/FeatureCard/FeatureCard.module.scss
