+++ components/templates/Home/SummarySection/SummarySection.module.scss
