+++ components/templates/Home/SummarySection/SummarySection.tsx
