+++ styles/globals.scss
@@ -1,6 +1,3 @@
-@import "/styles/mixin.scss";
-@import "/styles/variables.scss";
-
 html,
 body {
   padding: 0;
@@ -14,39 +11,7 @@ a {
 
 * {
   box-sizing: border-box;
-  
   font-family: "Noto Sans TC";
   font-style: normal;
   font-weight: normal;
-}
-
-button[data-direction] {
-  width: $width-arrow-d;
-  height: $height-arrow-d;
-  cursor: pointer;
-  border: none;
-  background: url('/images/common/icons/arrow_right.svg') no-repeat center;
-
-  @include large-tablet {
-    width: $width-arrow-lt;
-    height: $height-arrow-lt;
-    transform: translateX(-10px);
-  }
-
-  @include tablet {
-    width: $width-arrow-t;
-    height: $height-arrow-t;
-  }
-
-  @include mobile {
-    transform: unset;
-  }
-}
-
-button[data-direction='left'] { /* left button */
-  transform: rotateZ(180deg);
-
-  @include large-tablet {
-    transform: rotateZ(180deg) translateX(-10px);
-  }
 }
